merge new version

+++ priam/src/test/java/com/netflix/priam/backup/TestSnapshotStatusMgr.java
@@ -19,37 +19,31 @@ package com.netflix.priam.backup;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
-import com.netflix.priam.IConfiguration;
+import com.netflix.priam.config.IConfiguration;
 import com.netflix.priam.utils.DateUtil;
-import junit.framework.Assert;
+import java.io.File;
+import java.util.Date;
+import java.util.List;
 import org.joda.time.DateTime;
+import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.File;
-import java.util.Date;
-import java.util.List;
-
-/**
- * Created by aagrawal on 7/11/17.
- */
+/** Created by aagrawal on 7/11/17. */
 public class TestSnapshotStatusMgr {
     private static final Logger logger = LoggerFactory.getLogger(TestSnapshotStatusMgr.class);
 
-    private static Injector injector;
     private static IBackupStatusMgr backupStatusMgr;
-    private static IConfiguration configuration;
 
     @BeforeClass
     public static void setup() {
-        injector = Guice.createInjector(new BRTestModule());
-        //cleanup old saved file, if any
-        configuration = injector.getInstance(IConfiguration.class);
+        Injector injector = Guice.createInjector(new BRTestModule());
+        // cleanup old saved file, if any
+        IConfiguration configuration = injector.getInstance(IConfiguration.class);
         File f = new File(configuration.getBackupStatusFileLoc());
-        if (f.exists())
-            f.delete();
+        if (f.exists()) f.delete();
 
         backupStatusMgr = injector.getInstance(IBackupStatusMgr.class);
     }
@@ -106,6 +100,7 @@ public class TestSnapshotStatusMgr {
         Date startTime = DateUtil.getDate("19840101");
 
         for (int i = 0; i < noOfEntries; i++) {
+            assert startTime != null;
             Date time = new DateTime(startTime.getTime()).plusHours(i).toDate();
             BackupMetadata backupMetadata = new BackupMetadata("123", time);
             backupStatusMgr.start(backupMetadata);
@@ -116,11 +111,10 @@ public class TestSnapshotStatusMgr {
         Assert.assertEquals(noOfEntries, metadataList.size());
         logger.info(metadataList.toString());
 
-        //Ensure that list is always maintained from latest to eldest
+        // Ensure that list is always maintained from latest to eldest
         Date latest = null;
         for (BackupMetadata backupMetadata : metadataList) {
-            if (latest == null)
-                latest = backupMetadata.getStart();
+            if (latest == null) latest = backupMetadata.getStart();
             else {
                 Assert.assertTrue(backupMetadata.getStart().before(latest));
                 latest = backupMetadata.getStart();
@@ -134,6 +128,7 @@ public class TestSnapshotStatusMgr {
         Date startTime = DateUtil.getDate("19850101");
 
         for (int i = 0; i < noOfEntries; i++) {
+            assert startTime != null;
             Date time = new DateTime(startTime.getTime()).plusDays(i).toDate();
             BackupMetadata backupMetadata = new BackupMetadata("123", time);
             backupStatusMgr.start(backupMetadata);
@@ -141,7 +136,7 @@ public class TestSnapshotStatusMgr {
         }
 
         // Verify there is only capacity entries
-        Assert.assertEquals(backupStatusMgr.getCapacity(), backupStatusMgr.getAllSnapshotStatus().size());
+        Assert.assertEquals(
+                backupStatusMgr.getCapacity(), backupStatusMgr.getAllSnapshotStatus().size());
     }
-
 }
